feat: add V1 update decoder and ApplyUpdateV1
- Add UpdateDecoderV1.cs (DSDecoderV1 + UpdateDecoderV1) for Yjs V1 wire format
- Add YDoc.ApplyUpdateV1(Stream) and ApplyUpdateV1(byte[])
- Fix integer overflow in StreamDecodingExtensions.ReadAny (case 125)

These changes support reading V1-encoded updates produced by Yjs
encodeStateAsUpdate(), which is the default in the JS runtime.

+++ src/Ycs/lib0/StreamDecodingExtensions.cs
@@ -207,7 +207,7 @@ namespace Ycs
                     return BitConverter.ToSingle(fBytes);
 #endif // NETSTANDARD2_0
                 case 125: // integer
-                    return (int)stream.ReadVarInt().Value;
+                    return stream.ReadVarInt().Value;
                 case 126: // null
                 case 127: // undefined
                     return null;
